+++ packages/platforms/browser/lib/config.ts
@@ -1,12 +1,13 @@
 import {
-  isStringOrRegExpArray,
   isBoolean,
+  isStringOrRegExpArray,
   schema,
   type ConfigOption,
   type Configuration,
   type CoreSchema
 } from '@bugsnag/core-performance'
 import { isRoutingProvider, type RoutingProvider } from './routing-provider'
+import { defaultSendPageAttributes, isSendPageAttributes, type SendPageAttributes } from './send-page-attributes'
 
 export interface BrowserSchema extends CoreSchema {
   autoInstrumentFullPageLoads: ConfigOption<boolean>
@@ -15,6 +16,7 @@ export interface BrowserSchema extends CoreSchema {
   generateAnonymousId: ConfigOption<boolean>
   routingProvider: ConfigOption<RoutingProvider>
   settleIgnoreUrls: ConfigOption<Array<string | RegExp>>
+  sendPageAttributes: ConfigOption<SendPageAttributes>
 }
 
 export interface BrowserConfiguration extends Configuration {
@@ -24,6 +26,7 @@ export interface BrowserConfiguration extends Configuration {
   generateAnonymousId?: boolean
   routingProvider?: RoutingProvider
   settleIgnoreUrls?: Array<string | RegExp>
+  sendPageAttributes?: SendPageAttributes
 }
 
 export function createSchema (hostname: string, defaultRoutingProvider: RoutingProvider): BrowserSchema {
@@ -62,6 +65,11 @@ export function createSchema (hostname: string, defaultRoutingProvider: RoutingP
       defaultValue: [],
       message: 'should be an array of string|RegExp',
       validate: isStringOrRegExpArray
+    },
+    sendPageAttributes: {
+      defaultValue: defaultSendPageAttributes,
+      message: 'should be an object',
+      validate: isSendPageAttributes
     }
   }
 }